Add linuxArm64 and wasmJs targets (#166)
Kotest added support for wasm a whiel ago, but hasn't made a release since then. Rather than waiting any longer, we use a snapshot build of kotest for now. It's only a test dependency, so it doesn't affect the published mordant artifacts.

Wasm was a bit of a pain: you can't share much code between it and JS, because wasm introduces a `JsAny`, while js uses `dynamic`. The `js()` function also works differently between the two targets. Finally, wasm seems to translate `\r` to `\r\n` when printing, so I had to work around that for animations.