Support raw mode for posix in native image
The issue with `Expected Object but got Word` when compiling the native
image comes from the fact that the kotlin compiler insert null checks if
the used types are not nullable.

For example, if we declare tcgetattr as

```kotlin
@CFunction("tcgetattr")
external fun tcgetattr(fd: Int, termios: termios): Int
```

Every call to it will be compiled as

```kotlin
Intrinsics.checkNotNull(termios);
tcgetattr(..., termios);
```

It is possible to disable those checks with the following compiler
parameters:
-Xno-param-assertions, -Xno-call-assertions,-Xno-receiver-assertions.

But in this case, it is enough to simply mark the termios parameter as
nullable.