Prevent error with musl libc and GraalVM native image
When compiling a native image for linux with musl libc,
the following error was displayed:
`'struct termios' has no member named 'c_ispeed'`

On musl, it is named `__c_ispeed`

Since we don't really need those fields anyway,
we can remove them from the struct definition.

We then need to mark the struct definition as incomplete.
The documentation says:
If marked as incomplete, we will not try to determine the size of the struct.

`StackValue.get(structType)` is actually a shortcut for
`StackValue.get(SizeOf.get(structType))`

We now simply need to pass the explicit size in bytes.