Substitute TerminalInterfaceProvider
Before the new native image default `--strict-image-heap`,
we used `--initialize-at-build-time=com.github.ajalt.mordant.internal.MppImplKt`
so that the os was resolved at build time and classes of other OS were
not used.

With the new default, we need to specify every class that needs to be
initialized from the initialized class.

It's also easy to move or introduce a new class and not keep the
configuration synced.

We now use @Substitute + @Platforms to get the expected behaviour

Substituting TerminalInterfaceProviderJna isn't really needed, but it
prevents the jna classes from being included in the final binary.

As for the Ffm provider, we need to disable it as it requires a build
flag `-H:+ForeignAPISupport` + some configuration and isn't supported
with GraalVM 21.