Use consistent naming for LICENSE options (#252)
Hello, I am using this library in one of the projects and during the
license compliance check I noticed that its POM looks a bit strange - it
has 2 **license** blocks inside the
[licenses](https://repo1.maven.org/maven2/com/github/ajalt/mordant/mordant-jvm-jna-jvm/3.0.1/mordant-jvm-jna-jvm-3.0.1.pom)
block.

<details><summary>POM part</summary>
<p>

```xml
<licenses>
  <license>
    <name>Apache-2.0</name>
    <url>https://www.apache.org/licenses/LICENSE-2.0</url>
  </license>
  <license>
    <distribution>repo</distribution>
  </license>
</licenses>
```

</p>
</details> 

Right now there are two versions of license properties are used
(LICEN**S**E and LICEN**C**E) in gradle.properties. POM_LICENSE_DIST is
added as a separate license because of the following change in
gradle-maven-publish-plugin:
https://github.com/vanniktech/gradle-maven-publish-plugin/pull/272.

I renamed all the options to LICEN**S**E to match the naming in Gradle
API and the license itself.

Please, let me know If I have missed anything in the PR (update the
changelog for example)